Generate scope methods from parent models
Including relation methods modules from parent models does not work
properly since the method return type of the included scope methods are
still of the parent relation types.

A better solution is to collect all the scope methods from all the
parent models of the current model until we hit AR::Base or an abstract
model.