Rename `Gem` to `GemSpec`
The name `Gem` for the internal class to represent an internal class to
hold information about a gem was causing reference errors, and ended up
confusing top level `::Gem` with `Tapioca::Gemfile::Gem` in generated
RBI files.

`GemSpec` is a better and more descriptive name for it than `Gem`
anyway.